Fixing compliance with Issue 14682 "GetValue&lt;decimal&gt;() won't convert strings"
Issue Description:
ExcelWorksheet.GetValue&lt;decimal&gt;() always returns 0 for a cell that contains
a number stored as a string. The GetTypedValue() method has no case for this conversion,
and the final else returns a type's default value. GetTypedValue() looks more complicated
than it needs to be as most of its functionality can be achieved with something like:
<f>convertedValue = (T) Convert.ChangeType(v, typeof (T));</f>
where v is the value before conversion.

[patrick1] suggests that the GetTypedValue method should be simplified.
-------
Ted_250 replaced the return of default(T) with the suggested conversion call without replacing the existing special cases.