Excel Date1904 feature support
When EPPlus read an excel file created on macOS the dates are four years different compared to date showed in MS Excel.

The problem is well known and exists documentation about this issue (check it out references).

To fix in EPPlus we must read the value of this property (if exists)

xl/workbook.xml@date1904

and correct all date cells when reading and saving the sheet. Also if this setting it's modified must be reflected in whole workbook (like MS Excel does).


Test cases also was added for this scenario. Identical behavior as MS Excel with the setting enabled/disabled

REF.
http://support.microsoft.com/kb/103996
http://support.microsoft.com/kb/180162
http://blogmines.com/blog/2010/12/05/how-to-change-the-date-system-for-a-workbook-in-excel-2010/