When loading cells: flag cells containing array formula
Array formulas are not yet supported (according to the comment in `LoadCells (Xmlreader)` at line 1358 `else if (t == "array")`), but there is some basic support.

Unfortunately, however, at this point there is no way to recognize that a formula was originally an array formula. Solve this by setting the `CellFlags.ArrayFormula` flag for the relevant cell.