Don't create a big FileInfo object just to get the image file extension
FileInfo class is not permitted in AppDomain's that don't have code-level permission to access the file system.  In this case, we are just looking for the image's file extension, and we can use the Path.GetExtension method which just parses the string to do this, instead of actually creating an instance of the FileInfo object.