Replace all leading tabs with 4 spaces.
When viewing source files on the GitHub website, the inconsistent
use of tabs and spaces for indentation results in lines being
misaligned. The result is that it is harder to read and follow
code.

This change replaces all tabs in .cs files at the beginnings of lines
with 4 spaces. This seems to match the indentation convention currently
used by the authors and contributors.

Used command:

    find * -name '*.cs' -print0 | xargs -0 sed -i -e ':loop;s/^\( *\)\t/\1    /;tloop'

Fixes #303.