Use case-insensitive keys for the table column name dictionary
Excel chokes if a table contains column names that differ only in case, i.e. it displays the following message when the file is opened and renames the column:

"We found a problem with some content in 'YourFile.xlsx'. Do you want us to try to recover as much as we can? If you trust the source of this workbook, click Yes."