Fix bug in Series
It is possible that Series has no elements. That means that Series[0] cannot be accessed. It was the case in my Excel file that I was working on.

I only have ScatterCharts in my Excel file, thus, if the other charts also assume that Series always has elements, this will fail in those cases.